Split test_directive_productionlist
